[XLA] Allow dynamic shapes to propagate through sharding annotations without padding
Sharding dynamic shapes doesn't make sense, but since they can be mixed in a tuple with sharded shapes like infeed, we still need to specify a sharding for them. This triggers dynamic padder, as it now wants to pad the inputs to the sharding custom calls, which are purely symbolic. Don't let it do that.

PiperOrigin-RevId: 686668732